inject api and updater into tauri crate
