--- core/tauri-api/.license_template
@@ -1,1 +1,0 @@
-../../.license_template
