--- core/tauri-api/CHANGELOG.md
@@ -1,48 +1,0 @@
-# Changelog
-
-## [0.7.5]
-
--   `dirs` crate is unmaintained, now using `dirs-next` instead.
-    -   [82cda98](https://www.github.com/tauri-apps/tauri/commit/82cda98532975c6d4b1c93bf2f326173f39e0964) chore(tauri) `dirs` crate is unmaintained, use `dirst-next` instead ([#1057](https://www.github.com/tauri-apps/tauri/pull/1057)) on 2020-10-17
-    -   [72996be](https://www.github.com/tauri-apps/tauri/commit/72996be1bd3eb878c4cf30bfec79058071c26d7a) apply version updates ([#1024](https://www.github.com/tauri-apps/tauri/pull/1024)) on 2020-10-21
-
-## [0.7.4]
-
--   Bump all deps as noted in #975, #976, #977, #978, and #979.
-    -   [06dd75b](https://www.github.com/tauri-apps/tauri/commit/06dd75b68a15d388808c51ae2bf50554ae761d9d) chore: bump all js/rust deps ([#983](https://www.github.com/tauri-apps/tauri/pull/983)) on 2020-08-20
-
-## [0.7.3]
-
--   Use native dialog on `window.alert` and `window.confirm`.
-    Since every communication with the webview is asynchronous, the `window.confirm` returns a Promise resolving to a boolean value.
-        - [0245833](https://www.github.com/tauri-apps/tauri/commit/0245833bb56d5462a4e1249eb1e2f9f5e477592d) feat(tauri) make `window.alert` and `window.confirm` available, fix [#848](https://www.github.com/tauri-apps/tauri/pull/848) ([#854](https://www.github.com/tauri-apps/tauri/pull/854)) on 2020-07-18
-        - [dac0ae9](https://www.github.com/tauri-apps/tauri/commit/dac0ae976ea1b419ed5af078d00106b1476dee04) chore(changes) add tauri-api to JS dialogs changefile on 2020-07-19
-
-## [0.7.2]
-
--   Fixes Edge blank screen on Windows when running tauri dev (Tauri crashing window due to Edge reloading app because of missing Content-Type header).
-    -   Bumped due to a bump in tauri-api.
-    -   [fedee83](https://www.github.com/tauri-apps/tauri/commit/fedee835e36daa4363b91aabd43143e8033c9a5c) fix(tauri.js) windows Edge blank screen on tauri dev ([#808](https://www.github.com/tauri-apps/tauri/pull/808)) on 2020-07-11
-
-## [0.7.1]
-
--   Fixes the config reloading when tauri.conf.json changes.
-
-## [0.7.0]
-
--   The execute_promise and execute_promise_sync helpers now accepts any tauri::Result<T> where T: impl Serialize.
-    This means that you do not need to serialize your response manually or deal with String quotes anymore.
-    As part of this refactor, the event::emit function also supports impl Serialize instead of String.
--   readDir API refactor. Now returns path, name and children. 
-
-## [0.6.1]
-
--   Fixes the httpRequest headers usage. It now accepts Strings instead of serde_json::Value.
-
-## [0.6.0]
-
--   This adds HttpRequestBuilder, described at "alternatives you've considered" section in undefined.
--   Adds a command line interface option to tauri apps, configurable under tauri.conf.json > tauri > cli.
--   Fixes no-server mode not running on another machine due to fs::read_to_string usage instead of the include_str macro.
-    Build no longer fails when compiling without environment variables, now the app will show an error.
--   Adds desktop notifications API.
