--- core/tauri-api/Cargo.toml
@@ -1,54 +1,0 @@
-[package]
-name = "tauri-api"
-version = "0.7.5"
-authors = [
-  "Lucas Fernandes Gonçalves Nogueira <lucas@tauri.studio>",
-  "Daniel Thompson-Yvetot <denjell@sfosc.org>",
-  "Tensor Programming <tensordeveloper@gmail.com>"
-]
-categories = [ "gui", "os", "filesystem", "web-programming" ]
-license = "Apache-2.0 OR MIT"
-homepage = "https://tauri.studio"
-repository = "https://github.com/tauri-apps/tauri"
-description = "Make tiny, secure apps for all desktop platforms with Tauri"
-edition = "2018"
-exclude = [ "test/fixture/**" ]
-
-[dependencies]
-serde = { version = "1.0", features = [ "derive" ] }
-serde_json = { version = "1.0", features = [ "raw_value" ]}
-serde_repr = "0.1"
-dirs-next = "2.0.0"
-zip = "0.5.11"
-semver = "0.11"
-ignore = "^0.4.17"
-tempfile = "3"
-either = "1.6.1"
-tar = "0.4"
-flate2 = "1.0"
-thiserror = "1.0.24"
-rand = "0.8"
-rfd = "0.2.2"
-tinyfiledialogs = "3.3"
-reqwest = { version = "0.11", features = [ "json", "multipart" ] }
-bytes = { version = "1", features = ["serde"] }
-http = "0.2"
-tauri-utils = { version = "0.5", path = "../tauri-utils" }
-clap = { version = "=3.0.0-beta.2", optional = true }
-notify-rust = { version = "4.3.0", optional = true }
-once_cell = "1.7.2"
-tauri-hotkey = { git = "https://github.com/tauri-apps/tauri-hotkey-rs", branch = "dev", optional = true }
-open = "1.6.0"
-tokio = { version = "1.4", features = ["rt", "rt-multi-thread", "sync"] }
-shared_child = "0.3"
-os_pipe = "0.9"
-
-[dev-dependencies]
-quickcheck = "1.0.3"
-quickcheck_macros = "1.0.0"
-
-[features]
-build = [ "tauri-utils/build" ]
-cli = [ "clap" ]
-notification = [ "notify-rust" ]
-global-shortcut = [ "tauri-hotkey" ]
