+++ core/tauri-codegen/src/context.rs
@@ -6,14 +6,14 @@ use crate::embedded_assets::{EmbeddedAssets, EmbeddedAssetsError};
 use proc_macro2::TokenStream;
 use quote::quote;
 use std::path::PathBuf;
-use tauri_api::config::Config;
+use tauri_utils::config::Config;
 
 /// Necessary data needed by [`context_codegen`] to generate code for a Tauri application context.
 pub struct ContextData {
   pub dev: bool,
   pub config: Config,
   pub config_parent: PathBuf,
-  pub context_path: TokenStream,
+  pub root: TokenStream,
 }
 
 /// Build a `tauri::Context` for including in application code.
@@ -22,7 +22,7 @@ pub fn context_codegen(data: ContextData) -> Result<TokenStream, EmbeddedAssetsE
     dev,
     config,
     config_parent,
-    context_path,
+    root,
   } = data;
   let assets_path = if dev {
     // if dev_path is a dev server, we don't have any assets to embed
@@ -51,11 +51,11 @@ pub fn context_codegen(data: ContextData) -> Result<TokenStream, EmbeddedAssetsE
   };
 
   // double braces are purposeful to force the code into a block expression
-  Ok(quote!(#context_path {
+  Ok(quote!(#root::Context {
     config: #config,
     assets: #assets,
     default_window_icon: #default_window_icon,
-    package_info: ::tauri::api::PackageInfo {
+    package_info: #root::api::PackageInfo {
         name: env!("CARGO_PKG_NAME"),
         version: env!("CARGO_PKG_VERSION")
     }
