+++ core/tauri/.scripts/loop_qc.sh
