+++ core/tauri/Cargo.toml
@@ -14,11 +14,13 @@ description = "Make tiny, secure apps for all desktop platforms with Tauri"
 edition = "2018"
 exclude = [ "test/fixture/**" ]
 
+
 [package.metadata.docs.rs]
 features = [ "api-all" ]
 
+
 [dependencies]
-serde_json = "1.0"
+serde_json = { version = "1.0", features = [ "raw_value" ] }
 serde = { version = "1.0", features = [ "derive" ] }
 base64 = "0.13.0"
 tokio = { version = "1.4", features = ["rt", "rt-multi-thread", "sync"] }
@@ -26,14 +28,42 @@ futures = "0.3"
 uuid = { version = "0.8.2", features = [ "v4" ] }
 thiserror = "1.0.24"
 once_cell = "1.7.2"
-tauri-api = { version = "0.7.5", path = "../tauri-api" }
-tauri-updater = { version = "0.4.2", optional = true,  path = "../tauri-updater" }
 tauri-macros = { version = "0.1", path = "../tauri-macros" }
+tauri-utils = { version = "0.5", path = "../tauri-utils" }
 wry = "0.7"
 rand = "0.8"
 
+# api & updater
+reqwest = { version = "0.11", features = [ "json", "multipart" ] }
+tempfile = "3"
+semver = "0.11"
+
+# api
+serde_repr = "0.1"
+dirs-next = "2.0.0"
+zip = "0.5.11"
+ignore = "^0.4.17"
+either = "1.6.1"
+tar = "0.4"
+flate2 = "1.0"
+rfd = "0.2.2"
+tinyfiledialogs = "3.3"
+bytes = { version = "1", features = ["serde"] }
+http = "0.2"
+clap = { version = "=3.0.0-beta.2", optional = true }
+notify-rust = { version = "4.3.0", optional = true }
+tauri-hotkey = { git = "https://github.com/tauri-apps/tauri-hotkey-rs", branch = "dev", optional = true }
+open = "1.6.0"
+shared_child = "0.3"
+os_pipe = "0.9"
+
+# updater
+minisign-verify = "0.1.8"
+
+
 [build-dependencies]
 cfg_aliases = "0.1.1"
+
 
 [dev-dependencies]
 proptest = "1.0.0"
@@ -41,11 +71,20 @@ serde_json = "1.0"
 tauri = { path = ".", features = [ "api-all" ] }
 serde = { version = "1.0", features = [ "derive" ] }
 
+# api
+quickcheck = "1.0.3"
+quickcheck_macros = "1.0.0"
+
+# updater
+tokio-test ="*"
+mockito = "0.29"
+
+
 [features]
-cli = [ "tauri-api/cli" ]
+cli = [ "clap" ]
 custom-protocol = [ "tauri-macros/custom-protocol" ]
-api-all = [ "tauri-api/notification", "tauri-api/global-shortcut", "updater" ]
-updater = [ "tauri-updater" ]
+api-all = [ "notification-all", "global-shortcut-all", "updater" ]
+updater = [ "reqwest/default-tls" ]
 
 # FS
 fs-all = [ ]
@@ -80,7 +119,7 @@ http-all = [ ]
 http-request = [ ]
 
 # notification
-notification-all = [ "tauri-api/notification" ]
+notification-all = [ "notify-rust" ]
 
 # global shortcut
-global-shortcut-all = [ "tauri-api/global-shortcut" ]
+global-shortcut-all = [ "tauri-hotkey" ]
