+++ core/tauri/src/endpoints/cli.rs
@@ -21,7 +21,7 @@ impl Cmd {
       #[allow(unused_variables)]
       Self::CliMatches => {
         #[cfg(cli)]
-        return tauri_api::cli::get_matches(&cli_config)
+        return crate::api::cli::get_matches(&cli_config)
           .map_err(Into::into)
           .map(Into::into);
         #[cfg(not(cli))]
