+++ core/tauri/src/endpoints/file_system.rs
@@ -5,8 +5,8 @@
 use super::InvokeResponse;
 use crate::api::path::BaseDirectory;
 
+use crate::api::{dir, file, path::resolve_path};
 use serde::{Deserialize, Serialize};
-use tauri_api::{dir, file, path::resolve_path};
 
 use std::{fs, fs::File, io::Write, path::PathBuf};
