+++ core/tauri/src/updater/core.rs
@@ -2,10 +2,9 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-License-Identifier: MIT
 
-#[macro_use]
-pub mod error;
+use super::error::{Error, Result};
+use crate::api::{file::Extract, version};
 use base64::decode;
-pub use error::{Error, Result};
 use minisign_verify::{PublicKey, Signature};
 use reqwest::{self, header, StatusCode};
 use std::{
@@ -17,13 +16,12 @@ use std::{
   str::from_utf8,
   time::{Duration, SystemTime, UNIX_EPOCH},
 };
-use tauri_api::{file::Extract, version};
 
 #[cfg(not(target_os = "macos"))]
 use std::process::Command;
 
 #[cfg(target_os = "macos")]
-use tauri_api::file::Move;
+use crate::api::file::Move;
 
 #[cfg(target_os = "windows")]
 use std::process::exit;
@@ -175,6 +173,7 @@ impl<'a> UpdateBuilder<'a> {
     UpdateBuilder::default()
   }
 
+  #[allow(dead_code)]
   pub fn url(mut self, url: String) -> Self {
     self.urls.push(url);
     self
@@ -199,12 +198,14 @@ impl<'a> UpdateBuilder<'a> {
 
   /// Set the target (os)
   /// win32, win64, darwin and linux are currently supported
+  #[allow(dead_code)]
   pub fn target(mut self, target: &str) -> Self {
     self.target = Some(target.to_owned());
     self
   }
 
   /// Set the executable path
+  #[allow(dead_code)]
   pub fn executable_path<A: AsRef<Path>>(mut self, executable_path: A) -> Self {
     self.executable_path = Some(PathBuf::from(executable_path.as_ref()));
     self
