+++ core/tauri/test/api/test_binary
