+++ core/tauri/test/updater/fixture/archives/archive.tar.gz.badsig
