+++ core/tauri/test/updater/fixture/archives/archive.zip
