+++ core/tauri/test/updater/fixture/bad_signature/update.key.pub
