+++ core/tauri/test/updater/fixture/good_signature/update.key
